+++ src/functions/objects-array-to-object.spec.ts
@@ -1,8 +1,19 @@
 import { objectsArrayToObject } from './objects-array-to-object';
 
-describe('objectsArrayToObject', () => {
+describe('objectsArrayToObject DEPRECATED', () => {
+
+    let consoleSpy: jest.SpyInstance;
+
+    beforeEach(() => {
+        consoleSpy = jest.spyOn(console, 'warn').mockImplementation();
+    });
+
+    afterEach(() => {
+        consoleSpy.mockRestore();
+    });
+
     it('convert array to object', () => {
-        expect.assertions(1);
+        expect.assertions(3);
         const array = [
             {oneKey: 'oneValue', twoKey: 'twoValue'},
             {oneKey: 'threeValue', twoKey: 'fourValue'},
@@ -14,10 +25,14 @@ describe('objectsArrayToObject', () => {
             oneValue: {oneKey: 'oneValue', twoKey: 'twoValue'},
             threeValue: {oneKey: 'threeValue', twoKey: 'fourValue'},
         });
+        expect(consoleSpy).toHaveBeenCalledTimes(1);
+        expect(consoleSpy).toHaveBeenCalledWith(
+            'Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.'
+        );
     });
 
     it('convert array to object with number keys', () => {
-        expect.assertions(1);
+        expect.assertions(3);
         const array = [
             {1: 'oneValue', 2: 'twoValue'},
             {1: 'threeValue', 2: 'fourValue'},
@@ -28,6 +43,61 @@ describe('objectsArrayToObject', () => {
         expect(result).toStrictEqual({
             oneValue: {1: 'oneValue', 2: 'twoValue'},
             threeValue: {1: 'threeValue', 2: 'fourValue'},
+        });
+        expect(consoleSpy).toHaveBeenCalledTimes(1);
+        expect(consoleSpy).toHaveBeenCalledWith(
+            'Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.'
+        );
+    });
+});
+
+describe('objectsArrayToObject', () => {
+    it('convert array to object', () => {
+        expect.assertions(1);
+        const array = [
+            {oneKey: 'oneValue', twoKey: 'twoValue'},
+            {oneKey: 'threeValue', twoKey: 'fourValue'},
+        ];
+
+        const result = objectsArrayToObject(array, (item) => item.twoKey);
+
+        expect(result).toStrictEqual({
+            fourValue: {oneKey: 'threeValue', twoKey: 'fourValue'},
+            twoValue: {oneKey: 'oneValue', twoKey: 'twoValue'},
+        });
+    });
+
+    it('convert array to object with number keys', () => {
+        expect.assertions(2);
+        const array = [
+            {1: 'oneValue', 2: 'twoValue'},
+            {1: 'threeValue', 2: 'fourValue'},
+        ];
+
+        const result = objectsArrayToObject(array, (item) => item[1]);
+        expect(result).toStrictEqual({
+            oneValue: {1: 'oneValue', 2: 'twoValue'},
+            threeValue: {1: 'threeValue', 2: 'fourValue'},
+        });
+
+        const result2 = objectsArrayToObject(array, (item) => item['2']);
+        expect(result2).toStrictEqual({
+            fourValue: {1: 'threeValue', 2: 'fourValue'},
+            twoValue: {1: 'oneValue', 2: 'twoValue'},
+        });
+    });
+
+    it('convert array to object with number values', () => {
+        expect.assertions(1);
+        const array = [
+            {1: 1, 2: 2},
+            {1: 3, 2: 4},
+        ];
+
+        const result = objectsArrayToObject(array, (item) => item[1]);
+        expect(result).toStrictEqual({
+            1: {1: 1, 2: 2},
+            3: {1: 3, 2: 4},
         });
     });
 });