+++ src/functions/objects-array-to-object.ts
@@ -1,10 +1,18 @@
-export const objectsArrayToObject = <T>(array: T[], key: string) => {
+export const objectsArrayToObject = <T>(array: T[], attributeGetter: string | ((item: T) => string | number)) => {
 
-    const object: {[index: string]: T} = {};
+    let keyIsString = false;
+    if (typeof attributeGetter === 'string') {
+        // eslint-disable-next-line no-console
+        console.warn('Using objectsArrayToObject(T[], string) is deprecated, use a selector function as the second parameter.');
+        keyIsString = true;
+    }
+
+    const object: { [index: string]: T } = {};
 
     for (const item of array) {
         // @ts-ignore
-        object[item[key] as string] = item;
+        const key = keyIsString ? item[attributeGetter] as string : attributeGetter(item);
+        object[key] = item;
     }
 
     return object;