+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- objectsArrayToObject now takes a selector function, using a string as key is deprecated.
+
 ## [3.0.0] - 2020-01-24
 ### BREAKING CHANGES
 - Dropped support for Node.js 8.